# proto/gen/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /api/v1/memoArchive:export:
        post:
            tags:
                - MemoTransferService
            description: ExportMemoArchive exports notes and todos across every personal space.
            operationId: MemoTransferService_ExportMemoArchive
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ExportMemoArchiveRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ExportMemoArchiveResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/v1/memoArchive:import:
        post:
            tags:
                - MemoTransferService
            description: ImportMemoArchive validates an archive before importing its records.
            operationId: MemoTransferService_ImportMemoArchive
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ImportMemoArchiveRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ImportMemoArchiveResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
...
components:
    schemas:
...
        ExportMemoArchiveRequest:
            type: object
            properties: {}
        ExportMemoArchiveResponse:
            type: object
            properties:
                content:
                    type: string
                    description: ZIP bytes, limited to 128 MiB.
                    format: bytes
                filename:
                    type: string
                memoCount:
                    type: integer
                    format: int32
                attachmentCount:
                    type: integer
                    format: int32
                warnings:
                    type: array
                    items:
                        type: string
...
        GoogleProtobufAny:
            type: object
            properties:
                '@type':
                    type: string
                    description: The type of the serialized message.
            additionalProperties: true
            description: Contains an arbitrary serialized message along with a @type that describes the type of the serialized message.
...
        ImportMemoArchiveRequest:
            type: object
            properties:
                content:
                    type: string
                    description: ZIP bytes, limited to 128 MiB.
                    format: bytes
        ImportMemoArchiveResponse:
            type: object
            properties:
                imported:
                    type: integer
                    format: int32
                skipped:
                    type: integer
                    description: Existing memo IDs are never overwritten.
                    format: int32
                attachments:
                    type: integer
                    format: int32
                warnings:
                    type: array
                    items:
                        type: string
                errors:
                    type: array
                    items:
                        type: string
                    description: Per-record failures are returned with any completed import counts.
...
        Status:
            type: object
            properties:
                code:
                    type: integer
                    description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
                    format: int32
                message:
                    type: string
                    description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
                details:
                    type: array
                    items:
                        $ref: '#/components/schemas/GoogleProtobufAny'
                    description: A list of messages that carry the error details.  There is a common set of message types for APIs to use.
            description: 'The `Status` type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).'
...